requirements.txt
dkimpy --> 0.6.2
dnspython --> 1.15.0
iptools --> 0.6.1
